chore: migrate reactive-controllers pacakge to Swan
