+++ swan/dev-docs/MIGRATION-PRINCIPLES.md
@@ -44,6 +44,226 @@
     - Check for regressions in dependent components
     - Validate type correctness across the system
 
+For a detailed step-by-step guide on performing migrations with `git mv` while preserving file history, see the [Git Move Migration: Step-by-Step Guide](#git-move-migration-step-by-step-guide) section.
+
+## Git Move Migration: Step-by-Step Guide
+
+When migrating files from SWC to Swan using `git mv` to preserve history, follow this exact sequence of steps to avoid circular dependencies and build issues:
+
+### 1. Prepare the Environment
+
+Before performing any file moves, ensure your TypeScript configuration is properly set up:
+
+1. **Update the SWC Package's tsconfig.json**
+   Add a reference to the Swan project in the package you're migrating:
+
+    ```json
+    {
+        "extends": "../../tsconfig.json",
+        "compilerOptions": {
+            "composite": true,
+            "rootDir": "./"
+        },
+        "include": ["*.ts", "src/*.ts"],
+        "exclude": ["test/*.ts", "stories/*.ts"],
+        "references": [{ "path": "../../swan" }]
+    }
+    ```
+
+2. **Update Root tsconfig.json Path Mappings**
+   Ensure proper path mappings exist for the module category you're migrating:
+    ```json
+    "paths": {
+        "@spectrum-web-components/swan/base/*": ["../swan/src/base/*"],
+        "@spectrum-web-components/swan/shared/*": ["../swan/src/shared/*"],
+        "@spectrum-web-components/swan/reactive-controllers/*": ["../swan/src/reactive-controllers/*"],
+        "@spectrum-web-components/swan/*": ["../swan/src/components/*"]
+    }
+    ```
+
+### 2. Migration Process
+
+For each file you want to migrate, follow these steps in order:
+
+1. **Create a Backup of the Original File**
+
+    ```bash
+    cp tools/[package]/src/[file].ts tools/[package]/src/[file].ts.orig
+    ```
+
+2. **Perform Git Move Operation**
+
+    ```bash
+    git mv tools/[package]/src/[file].ts swan/src/[package-path]/[file].ts
+    ```
+
+3. **Restore the Original File**
+
+    ```bash
+    mv tools/[package]/src/[file].ts.orig tools/[package]/src/[file].ts
+    ```
+
+4. **Update Swan Implementation**
+
+    - Open the moved file in Swan
+    - Update all imports to use proper Swan paths
+    - Fix any cross-module references
+
+5. **Configure Swan's package.json**
+   Add export entries for each migrated file:
+
+    ```json
+    "./[package-path]/[file].js": {
+        "types": "./dist/[package-path]/[file].d.ts",
+        "development": "./dist/[package-path]/[file].dev.js",
+        "default": "./dist/[package-path]/[file].js"
+    }
+    ```
+
+6. **Build Swan**
+
+    ```bash
+    cd swan && yarn build
+    ```
+
+7. **Replace Original with Shim**
+   Replace the original file content with a shim that re-exports from Swan:
+
+    ```typescript
+    /**
+     * ⚠️ IMPORTANT: SOURCE OF TRUTH MOVED ⚠️
+     *
+     * The authoritative implementation is now in:
+     * swan/src/[package-path]/[file].ts
+     *
+     * This file is maintained for backward compatibility ONLY.
+     * DO NOT modify this file directly; instead make changes in Swan.
+     */
+
+    export * from '@spectrum-web-components/swan/[package-path]/[file].js';
+    ```
+
+8. **Build and Test**
+
+    ```bash
+    yarn build
+    yarn test:focus [package]
+    ```
+
+9. **Remove SWC Package Dependency from Swan**
+    - Once all files from a package have been migrated and tests are passing, remove that package from Swan's dependencies:
+    ```diff
+    // In swan/package.json
+    "dependencies": {
+        "@internationalized/number": "^3.2.0",
+        "@lit-labs/observers": "^2.0.2",
+        "colorjs.io": "^0.5.2",
+    -   "@spectrum-web-components/reactive-controllers": "workspace:*",
+        "focus-visible": "^5.2.0",
+        "lit": "^2.5.0 || ^3.1.3"
+    }
+    ```
+    - Update any imports in Swan to use relative paths instead of package imports
+    - Build and test again to ensure there are no circular dependencies
+
+### 3. Why This Order Matters
+
+Following this specific order is crucial for several reasons:
+
+1. **Preserving Git History**: Using `git mv` maintains the complete history of the file
+2. **Avoiding Build Breaks**: Keeping the original file while preparing Swan prevents broken builds
+3. **Preventing Circular Dependencies**: Building Swan before creating shims ensures exports exist
+4. **TypeScript Project References**: The tsconfig changes establish proper project boundaries
+5. **Clean Dependency Management**: Removing SWC package dependencies from Swan after migration prevents circular dependencies and ensures proper encapsulation
+
+This carefully orchestrated sequence allows for gradual migration without disrupting the build process or creating circular dependencies. Skipping steps or performing them out of order can lead to difficult-to-debug build or runtime errors.
+
+### 4. Common Issues and Solutions
+
+-   **Error: Cannot find module**: Ensure Swan's package.json exports are correctly configured
+-   **TypeScript resolution errors**: Verify tsconfig path mappings are properly set up
+-   **404 errors during testing**: Check that Swan build completed successfully before creating shims
+-   **Circular dependencies**: Ensure you follow the steps in the exact order specified
+-   **Build script errors**: Some files with specific naming patterns may require special handling in build scripts
+
+### 5. Real-World Example: RovingTabindex Controller
+
+Here's a concrete example of migrating the RovingTabindex controller from reactive-controllers:
+
+**Step 1-3: Back up, move, and restore**
+
+```bash
+# Back up the original file
+cp tools/reactive-controllers/src/RovingTabindex.ts tools/reactive-controllers/src/RovingTabindex.ts.orig
+
+# Perform the git move
+git mv tools/reactive-controllers/src/RovingTabindex.ts swan/src/reactive-controllers/RovingTabindex.ts
+
+# Restore the original in its location
+mv tools/reactive-controllers/src/RovingTabindex.ts.orig tools/reactive-controllers/src/RovingTabindex.ts
+```
+
+**Step 4: Update imports in Swan implementation**
+
+Before:
+
+```typescript
+import { FocusGroupConfig, FocusGroupController } from './FocusGroup';
+```
+
+After:
+
+```typescript
+import { FocusGroupConfig, FocusGroupController } from './FocusGroup.js';
+```
+
+**Step 5: Configure Swan's package.json**
+
+```json
+"./reactive-controllers/RovingTabindex.js": {
+    "types": "./dist/reactive-controllers/RovingTabindex.d.ts",
+    "development": "./dist/reactive-controllers/RovingTabindex.dev.js",
+    "default": "./dist/reactive-controllers/RovingTabindex.js"
+}
+```
+
+**Step 6-7: Build Swan and create shim**
+
+After Swan is built, replace the original file with a shim:
+
+```typescript
+/*
+Copyright 2020 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
+
+/**
+ * ⚠️ IMPORTANT: SOURCE OF TRUTH MOVED ⚠️
+ *
+ * The authoritative implementation is now in:
+ * swan/src/reactive-controllers/RovingTabindex.ts
+ *
+ * This file is maintained for backward compatibility ONLY.
+ * DO NOT modify this file directly; instead make changes in Swan.
+ */
+
+export * from '@spectrum-web-components/swan/reactive-controllers/RovingTabindex.js';
+```
+
+**Step 8: Build and test**
+
+```bash
+yarn build
+yarn test:focus reactive-controllers
+```
+
 ## Shim Pattern
 
 ```typescript
