+++ tools/reactive-controllers/src/MatchMedia.ts
@@ -9,41 +9,15 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import type { ReactiveController, ReactiveElement } from 'lit';
 
-export const DARK_MODE = '(prefers-color-scheme: dark)';
-export const IS_MOBILE =
-    '(max-width: 743px) and (hover: none) and (pointer: coarse)';
+/**
+ * ⚠️ IMPORTANT: SOURCE OF TRUTH MOVED ⚠️
+ *
+ * The authoritative implementation is now in:
+ * swan/src/reactive-controllers/MatchMedia.ts
+ *
+ * This file is maintained for backward compatibility ONLY.
+ * DO NOT modify this file directly; instead make changes in Swan.
+ */
 
-export class MatchMediaController implements ReactiveController {
-    key = Symbol('match-media-key');
-
-    matches = false;
-
-    protected host: ReactiveElement;
-
-    protected media: MediaQueryList;
-
-    constructor(host: ReactiveElement, query: string) {
-        this.host = host;
-        this.host.addController(this);
-        this.media = window.matchMedia(query);
-        this.matches = this.media.matches;
-        this.onChange = this.onChange.bind(this);
-        host.addController(this);
-    }
-
-    public hostConnected(): void {
-        this.media?.addEventListener('change', this.onChange);
-    }
-
-    public hostDisconnected(): void {
-        this.media?.removeEventListener('change', this.onChange);
-    }
-
-    protected onChange(event: MediaQueryListEvent): void {
-        if (this.matches === event.matches) return;
-        this.matches = event.matches;
-        this.host.requestUpdate(this.key, !this.matches);
-    }
-}
+export * from '@spectrum-web-components/swan/reactive-controllers/MatchMedia.js';
